+++ src/engine/circuit_engine.cpp
@@ -20,9 +20,9 @@ std::vector<Qubit> CircuitEngine::qft(const std::vector<Qubit> j) {
   std::vector<Qubit> result(j.size());
 
   auto last_index = result.size() - 1;
-  for (int i = 0; i < j.size(); i++) {
+  for (size_t i = 0; i < j.size(); i++) {
     auto j_k = GateEngine::hadamard(j[i]);
-    for (int k_next = i + 1; k_next < j.size(); k_next++) {
+    for (size_t k_next = i + 1; k_next < j.size(); k_next++) {
       j_k = GateEngine::controlled_u(*j_k, j[k_next],
                                      *GateEngine::r_k(k_next + 1));
     }
@@ -41,7 +41,7 @@ std::vector<Qubit> CircuitEngine::inverse_qft(const std::vector<Qubit> k) {
   auto last_index = swapped.size() - 1;
   for (int i = last_index; i >= 0; i--) {
     auto k_i = swapped[i];
-    for (int j = last_index; j > i; j--) {
+    for (size_t j = last_index; j > i; j--) {
       k_i = *GateEngine::controlled_u(k_i, result[j],
                                       *GateEngine::r_k(j + 1, true));
     }
